+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.0.3] - 2018-11-03
+### Added
+- Added custom `@DOMOTICZON` and `@DOMOTICZOFF` GCODE commands.
+- Bypass previously delayed power off trigger if `Warn While Printing` option is enabled.
+
 ## [0.0.2] - 2018-07-07
 ### Added
 - User authentication to Domoticz server.
@@ -16,5 +21,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release.
 
+[0.0.3]: https://github.com/jneilliii/OctoPrint-Domoticz/tree/0.0.3
 [0.0.2]: https://github.com/jneilliii/OctoPrint-Domoticz/tree/0.0.2
 [0.0.1]: https://github.com/jneilliii/OctoPrint-Domoticz/tree/0.0.1